swap, addliquidity using uniswap/sushiswap

+++ src/interfaces/IFactoryStorage.sol
@@ -1,15 +1,16 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
+// SPDX-License-Identifier: GPL-3.0
 
 pragma solidity 0.7.5;
 
+/// @title IFactoryStorage Interface
 interface IFactoryStorage {
     function pushBond(
-        address _payoutToken, 
-        address _principleToken, 
-        address _customTreasury, 
-        address _customBond, 
-        address _initialOwner, 
-        uint[] calldata _tierCeilings, 
-        uint[] calldata _fees
-    ) external returns(address _treasury, address _bond);
+        address _payoutToken,
+        address _principleToken,
+        address _customTreasury,
+        address _customBond,
+        address _initialOwner,
+        uint256[] calldata _tierCeilings,
+        uint256[] calldata _fees
+    ) external returns (address _treasury, address _bond);
 }
